feat(Users): create Users/ model
Working on making the db modular
