+++ Network/Database/Database.hpp
@@ -11,17 +11,12 @@
 #include <sqlite3.h>
 #include <memory>
 #include <vector>
-// #include "Scores.hpp"
-    
 
+
+class Scores;
 
 class Database {
 
-    struct userSettings {
-        int res_width;
-        int res_height;
-        float daltonian_mode;
-    };
 
     public:
 
@@ -39,81 +34,12 @@ class Database {
         ~Database() = default;
 
         /**
-         * @brief Check if a user id exists
-         *
-         * @param id
-         * @return true
-         * @return false
-         */
-        bool userIdExists(int id);
-
-        /**
-         * @brief Register a user
-         *
-         * @param username
-         * @param password
-         * @return int The id of the user.
-         */
-        int registerUser(std::string username, std::string password);
-
-        /**
-         * @brief Login a user
-         *
-         * @param username
-         * @param password
-         * @return int The id of the user.
-         */
-        int loginUser(std::string username, std::string password);
-
-        /**
          * @brief Create the required tables for the database.
          *
          * @return true
          * @return false
          */
-        bool createTables();
-
-        /**
-         * @brief Hash a password
-         *
-         * @param password
-         * @return std::string
-         */
-        std::string hashPassword(std::string password);
-
-        /**
-         * @brief Check if a hashed pwd is another clear pwd.
-         *
-         * @param password
-         * @return bool
-         */
-        bool verifyPassword(std::string password, std::string hash);
-
-        /**
-         * @brief Get the user data
-         *
-         * @param id
-         * @return userSettings
-         */
-        Database::userSettings getUserSettings(int id);
-
-        /**
-         * @brief Set the User Settings object
-         *
-         * @param id
-         * @param res_width
-         * @param res_height
-         * @param daltonian_mode
-         */
-        bool setUserSettings(int id, int res_width, int res_height, float daltonian_mode);
-
-        /**
-         * @brief Get the username by id
-         *
-         * @param id
-         * @return std::string
-         */
-        std::string getUsernameById(int id);
+        // bool createTables();
 
         /**
          * @brief Prepare a statement
@@ -125,5 +51,4 @@ class Database {
 
         std::shared_ptr<sqlite3>_db = nullptr; // Database.
         std::shared_ptr<int> _rc = std::make_shared<int>(0); // Return code of db.
-        // Scores scores = Scores(this); //The scores model
 };
